+++ src/FactsFactory.php
@@ -36,9 +36,9 @@ class FactsFactory implements FactsFactoryInterface {
 			'modelOptions'			  => array(),
 			'entityThreshold'		  => 100, // min numbers of entity that trigger error resilence computation.
 			'resilienceToErrors' 	  => 0.3, // if more than 30% of error throws a TooManyErrorException
-			'resilienceToInsanes'	  => 0.9, // if more than 90% of unacceptable data throws a TooManyErrorException
-		    'source' 			  	  => null,
+		    'resilienceToInsanes'	  => 0.9, // if more than 90% of unacceptable data throws a TooManyErrorException
 		    'documentURL' 	          => null,	// used if printing metadata to specify source file URL use empty string for <>
+		    'source' 			  	  => null,
 			'datamapper'			  => function($rawdata){return $rawdata;},
 			'dataCleaner' 		  	  => get_class().'::REMOVE_EMPTY',
 			'factsErrorDetector' 	  => get_class().'::NOT_EMPTY_FACTS',
@@ -79,6 +79,7 @@ class FactsFactory implements FactsFactoryInterface {
 	    return array_filter($a);
 	}
 	
+	
 	/**
 	 * a default for dataValidator callback  
 	 */
@@ -92,8 +93,9 @@ class FactsFactory implements FactsFactoryInterface {
 	 * create facts from rawdata. Please nothe that null facts does not means always an error (i.e. no facts is a fact).
 	 * if you do not want empty facts use dataValidator
 	 */
-	public function factualize($rawData)
+	public function factualize($rawData,  object $globalStorage=null)
 	{
+	    
 		$rawdataSanitizer = $this->profile['rawdataSanitizer'];
 		$validRawData = $rawdataSanitizer($rawData);
 		$this->counter['entity']++;
@@ -103,7 +105,7 @@ class FactsFactory implements FactsFactoryInterface {
 			$dataCleaner = $this->profile['dataCleaner'];
 			$factsErrorDetector = $this->profile['factsErrorDetector'];
 			$data =$dataCleaner($datamapper($validRawData));
-			$facts = call_user_func($this->modelClass.'::fromArray',$data,$this->profile['modelOptions']);
+			$facts = call_user_func($this->modelClass.'::fromArray', $data, $this->profile['modelOptions'], $globalStorage);
 			$this->counter['triple'] += $facts->getTripleCount();
 			if($error=$factsErrorDetector($facts)){
 				$this->counter['error']++;
@@ -143,8 +145,7 @@ class FactsFactory implements FactsFactoryInterface {
 	        if(!empty($this->profile['source'])){
 	            $metadata .= "<{$this->profile['documentURL']}> dct:source <{$this->profile['source']}>. \n";
 	        }
-	        $metadata .= "[] a void:Dataset; void:datadump <{$this->profile['documentURL']}>;void:triples {$this->counter['triple']} ;void:entities {$this->counter['entity']}.\n";
-	        
+	        $metadata .= "[] a void:Dataset; dct:format \"text/turtle\" ; void:datadump <{$this->profile['documentURL']}>;void:triples {$this->counter['triple']} ;void:entities {$this->counter['entity']}.\n";	        
 	    }
 	    $base = $this->profile['modelOptions']['base']['default']?? null;
 	    return call_user_func($this->modelClass.'::getTurtleHeader',$base) . $metadata;